% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_qqplot.R
\name{create_df_qq}
\alias{create_df_qq}
\title{Create a table used for QQ-plotting}
\usage{
create_df_qq(ps, ci = 0.95)
}
\arguments{
\item{ps}{- vector with p-values; if a named vector is given, the creation
of Q-values will be performed in groups based on the levels of names}

\item{ci}{- confidence interval (default: 95%)}
}
\value{
tibble (data.frame) with following columns:
 \itemize{
   \item observed - -log10(observed p-values)
   \item expected - -log10(expected)
   \item clower - lower limit of CI
   \item cupper - upper limit of CI
   \item group - name of the group (if any)
 }
}
\description{
Create a table used for QQ-plotting
}
