% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_qqplot.R
\name{gg_qqplot}
\alias{gg_qqplot}
\title{Create a quantile-quantile plot with ggplot2.}
\usage{
gg_qqplot(ps, ci = 0.95, eq.axes = TRUE)
}
\arguments{
\item{ps}{Vector of p-values; if named, the names give groups.}

\item{ci}{Size of the confidence interval, 95\% by default.}

\item{eq.axes}{Logical: should the axes show equal range? Default: TRUE.}
}
\value{
A ggplot2 plot.
}
\description{
Assumptions:
  - Expected P values are uniformly distributed.
  - Confidence intervals assume independence between tests.
    We expect deviations past the confidence intervals if the tests are
    not independent.
    For example, in a genome-wide association study, the genotype at any
    position is correlated to nearby positions. Tests of nearby genotypes
    will result in similar test statistics.
}
\details{
Adapted by Julia Romanowska to include data grouping
}
\examples{
library(ggplot2)
gg_qqplot(runif(1e2)) + theme_grey(base_size = 24)

}
