% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabGenes.R
\name{grabGenes}
\alias{grabGenes}
\title{Find genes positioned in the given region; uses karyoploteR
 TxDb.Hsapiens.UCSC.hg19.knownGene, org.Hs.eg.db, and regioneR}
\usage{
grabGenes(positions, asGRangesList = TRUE)
}
\arguments{
\item{positions}{- data.frame with 'chromosome_name", 'start', and 'end'
columns (all other columns will be ignored)}

\item{asGRangesList}{- whether to return the results as GRanges (check
regioneR::GRanges for help) - boolean, defaults to FALSE}
}
\description{
Find genes positioned in the given region; uses karyoploteR
 TxDb.Hsapiens.UCSC.hg19.knownGene, org.Hs.eg.db, and regioneR
}
