% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabGenesEnsembl.R
\name{grabGenesEnsembl}
\alias{grabGenesEnsembl}
\title{grab gene regions from ensembl - using biomaRt}
\usage{
grabGenesEnsembl(
  positions,
  attribs,
  filters,
  asGRanges = FALSE,
  genome_ver = NULL
)
}
\arguments{
\item{positions}{- positions of markers for which to search the
regions in the mart, given as a data.frame with columns named as
filters that are given below}

\item{attribs}{- which attributes to fetch - check description in biomaRt::getBM}

\item{filters}{- which filters to use - check description in biomaRt::getBM}

\item{asGRanges}{- whether to return the results as GRanges (check
regioneR::GRanges for help) - boolean, defaults to FALSE}

\item{mart}{- which database to use - check description in biomaRt::getBM}

\item{genom_ver}{- which genome version? default: NULL - newest version}
}
\description{
grab gene regions from ensembl - using biomaRt
}
