% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabPositions.R
\name{grabGenesPositions}
\alias{grabGenesPositions}
\title{Fetching genes positions based on their names or IDs}
\usage{
grabGenesPositions(
  gene_symbols,
  filter_name = "hgnc_symbol",
  attribs = c("chromosome_name", "start_position", "end_position", "hgnc_symbol"),
  asGRanges = FALSE,
  genome_ver = NULL
)
}
\arguments{
\item{gene_symbols}{- character vector with names or IDs of the genes}

\item{filter_name}{- which filter is used? Check `listFilters` function in
the {biomaRt} package for help (default: "hgnc_symbol")}

\item{attribs}{- which attributes do you want to fetch? Check `listAttributes`
in {biomaRt} package; default: `c('chromosome_name', 'start_position',
'end_position', 'hgnc_symbol')`}

\item{asGRanges}{- should the result be returned in GRanges format? (default:
FALSE)}

\item{genome_ver}{- by default, function will fetch information from the
latest genome release; if any specific version is required, please provide
it here (e.g., "37" will give GRCh37)}
}
\value{
data.frame or GRanges object with the fetched information; columns 
  are named as the given attributes
}
\description{
Fetching genes positions based on their names or IDs
}
