% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabRegulRegions.R
\name{grabRegulRegions}
\alias{grabRegulRegions}
\title{Grab regulatory regions from H.sapiens genome on ensembl - using biomaRt}
\usage{
grabRegulRegions(positions, mart, attribs, filters, asGRanges = FALSE)
}
\arguments{
\item{positions}{- positions of markers for which to search the
regions in the mart, given as a data.frame with columns named as
filters that are given below}

\item{mart}{- which database to use - check description in biomaRt::getBM}

\item{attribs}{- which attributes to fetch - check description in biomaRt::getBM}

\item{filters}{- which filters to use - check description in biomaRt::getBM}

\item{asGRanges}{- whether to return the results as GRanges (check
regioneR::GRanges for help) - boolean, defaults to FALSE}
}
\description{
Grab regulatory regions from H.sapiens genome on ensembl - using biomaRt
}
