% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabPositions.R
\name{grabSNPpositions}
\alias{grabSNPpositions}
\title{Fetching SNP positions based on their names}
\usage{
grabSNPpositions(snp_names, attribs, asGRanges = FALSE, genome_ver = NULL)
}
\arguments{
\item{snp_names}{- character vector with names of the SNPs}

\item{attribs}{- which attributes do you want to fetch? Check `listAttributes`
in {biomaRt} package}

\item{asGRanges}{- should the result be returned in GRanges format? (default:
FALSE)}

\item{genome_ver}{- by default, function will fetch information from the
latest genome release; if any specific version is required, please provide
it here (e.g., "37" will give GRCh37)}
}
\value{
data.frame or GRanges object with the fetched information; columns 
  are named as the given attributes
}
\description{
Fetching SNP positions based on their names
}
